[GSP-7], [GSP-8], [GSP-9]
