+++ src/main/java/com/gridstudentpractice/chatservice/message/controller/MessageController.java
@@ -1,27 +1,22 @@
-package com.gridstudentpractice.chatservice.messages;
+package com.gridstudentpractice.chatservice.message.controller;
 
+import com.gridstudentpractice.chatservice.message.model.Message;
+import com.gridstudentpractice.chatservice.message.service.MessageService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.WebDataBinder;
-import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import javax.validation.Valid;
-import java.text.SimpleDateFormat;
-import java.time.LocalDateTime;
-import java.util.List;
 
 @Controller
 @RequestMapping("/chat")
 public class MessageController {
 
     @Autowired
-    private InMemoryMessageService messageService;
+    private MessageService messageService;
 
 //    @InitBinder
 //    public void initBinder(WebDataBinder binder) {
@@ -31,7 +26,7 @@ public class MessageController {
 //    }
 
     @RequestMapping(method = RequestMethod.GET)
-    public String getAllMessages(Model model) {
+    public String getMessages(Model model) {
         model.addAttribute("message", new Message());
         return "page";
     }
