+++ src/main/java/com/gridstudentpractice/chatservice/message/model/Message.java
@@ -1,4 +1,4 @@
-package com.gridstudentpractice.chatservice.messages;
+package com.gridstudentpractice.chatservice.message.model;
 
 import java.time.LocalDateTime;
