+++ src/main/java/com/gridstudentpractice/chatservice/message/service/MessageService.java
@@ -1,0 +1,10 @@
+package com.gridstudentpractice.chatservice.message.service;
+
+import com.gridstudentpractice.chatservice.message.model.Message;
+
+import java.util.List;
+
+public interface MessageService {
+    void sendMessage(Message message);
+    List<Message> getMessages();
+}
