--- src/main/java/com/gridstudentpractice/chatservice/messages/MessageServiceInterface.java
@@ -1,8 +1,0 @@
-package com.gridstudentpractice.chatservice.messages;
-
-import java.util.List;
-
-public interface MessageServiceInterface {
-    void sendMessage(Message message);
-    List<Message> getMessages();
-}
